Mexican 2020 states dataset

A data.frame containing population estimates for all the Mexican states in 2020

Usage

df_mxstate_2020

Format

An object of class data.frame with 32 rows and 11 columns.

Value

A data.frame

Details

region

INEGI code of the state

state_name

short state name (e.g. Coahuila)

state_name_official

Official state name (e.g. Coahuila de Zaragoza)

state_abbr

state abbreviation

state_abbr_official

official state abbreviation (it can be awkward to use Chis for Chiapas) according to the INEGI.

year

2015, the year of the Conteo from which the data is sourced

pop

total state population according to the Censo 2020

pop_male

male population according to the Censo 2020

pop_female

female population according to the Censo 2020

afromexican

afromexican population according to the Censo 2020

indigenous_language

Number of persons who speak an indigenous language according to the Censo 2020

References

Population estimates taken from the Censo 2020.

Examples

data("df_mxstate_2020")
head(df_mxstate_2020)
#>   region          state_name  state_name_official state_abbr
#> 1     01      Aguascalientes       Aguascalientes        AGS
#> 2     02     Baja California      Baja California         BC
#> 3     03 Baja California Sur  Baja California Sur        BCS
#> 4     04            Campeche             Campeche       CAMP
#> 5     05            Coahuila Coahuila de Zaragoza       COAH
#> 6     06              Colima               Colima        COL
#>   state_abbr_official year     pop pop_male pop_female afromexican
#> 1                Ags. 2020 1425607   696683     728924       22425
#> 2                  BC 2020 3769020  1900589    1868431       64362
#> 3                 BCS 2020  798447   405879     392568       26330
#> 4               Camp. 2020  928363   456939     471424       19319
#> 5               Coah. 2020 3146771  1563669    1583102       45976
#> 6                Col. 2020  731391   360622     370769       13574
#>   indigenous_language
#> 1                2539
#> 2               49130
#> 3               13581
#> 4               91801
#> 5                5527
#> 6                5210